Client: Fix "failed to allocate new peer" after leaving a crashed server
It happened because the disconnection process was waiting for an answer from the server before timing out after ~30s, if you were to reconnect before the previous connection times up the previous peer was still allocated.
Raising the peer count to 5 should be more than enough to handle this case.